+++ CppEducation/Alghoritms/InsertionSorter.cpp
@@ -2,7 +2,7 @@
 
 std::vector<int32_t> InsertionSorter::sort(std::vector<int32_t>& arr)
 {
-	for (int32_t i = 1; i < arr.size(); i++)
+	for (size_t i = 1; i < arr.size(); i++)
 	{
 		const auto curElement = arr[i];
 		int32_t j = i;
